main app rebuild
